refactor: rename JWT handler to middleware for user authentication and context management
