+++ cmd/backend/main.go
@@ -125,10 +125,10 @@ func main() {
 	// Handler
 	// exampleHandler := example.NewHandler(logger, validator, problemWriter, exampleService)
 	authHandler := auth.NewHandler(cfg, logger, validator, problemWriter, userService, jwtService, jwtService)
-	userHandler := user.NewHandler(logger, validator, problemWriter, userService)
 
 	// Middleware
 	traceMiddleware := trace.NewMiddleware(logger, cfg.Debug)
+	jwtMiddleware := jwt.NewMiddleware(logger, validator, problemWriter, jwtService)
 
 	// Basic Middleware (Tracing and Recovery)
 	basicMiddleware := middleware.NewSet(traceMiddleware.RecoverMiddleware)
